Extract logic from the OMGCaller to OMGCallbackWrapper
